# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TKK Control API",
        "description": "API documentation for TKK-Control",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/participants": {
            "get": {
                "tags": [
                    "Participants"
                ],
                "summary": "Get all participants",
                "operationId": "b56cc0b1479de99c053898dea57d2a59",
                "responses": {
                    "200": {
                        "description": "List of participants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Participants"
                ],
                "summary": "Create a new participant",
                "operationId": "700a91d1e48b221e6d27eb86f50a4dc7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "12345678A"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+34123456789"
                                    },
                                    "emergency_phone": {
                                        "type": "string",
                                        "example": "+34987654321"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "not_presented",
                                            "presented",
                                            "abandoned",
                                            "finished"
                                        ],
                                        "example": "presented"
                                    },
                                    "lunch_sandwich": {
                                        "type": "string",
                                        "example": true
                                    },
                                    "dinner_sandwich": {
                                        "type": "string",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Participant created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/participants/{id}": {
            "get": {
                "tags": [
                    "Participants"
                ],
                "summary": "Get a participant by ID",
                "operationId": "022e40e8979c7d7a7f156a8c80aeaa5f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Participant ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Participant details with related checks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Participant not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Participants"
                ],
                "summary": "Update a participant",
                "operationId": "c42492b70d7706860bae3c7f4966cfa4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Participant ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Jane"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Smith"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "not_presented",
                                            "presented",
                                            "abandoned",
                                            "finished"
                                        ],
                                        "example": "finished"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Participant updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Participant not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Participants"
                ],
                "summary": "Delete a participant",
                "operationId": "2a2003f3c1c8a6b6164344f3bd5085ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Participant ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Participant deleted"
                    },
                    "404": {
                        "description": "Participant not found"
                    }
                }
            }
        },
        "/api/admin/participants/import": {
            "post": {
                "tags": [
                    "Participants"
                ],
                "summary": "Import participants from CSV",
                "description": "Upload a CSV file to import multiple participants. Existing checks are truncated.",
                "operationId": "8ee236debdac3c7ead030924002d73a3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "csv"
                                ],
                                "properties": {
                                    "csv": {
                                        "description": "CSV file with participants",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import result with added count and errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "added": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid file"
                    }
                }
            }
        }
    },
...
}